feat : test Score Entity, refactor : RequestMentee >Participant

+++ mztalk-mentor-service/src/main/java/com/mztalk/mentor/domain/Entity/Score.java
@@ -17,7 +17,7 @@ public class Score extends BaseTimeEntity{
     @Column(name="score_id")
     private Long id;
 
-    private int count;
+    private Double count;
 
     @Column(nullable = true)
     private String content;
